1.18.2 support and fixes (#17)
* Add: 1.18 support

* Update to 1.18.1

* Fix insecure link

* Leave TODO

* Fix some general stuff

* Convert to UTF8 again

* Workaround for conditional crash on world unload

* Fixed attempt to process `AreaBlockEffects` before connection;

* Maybe fixed https://github.com/OreCruncher/DynamicSurroundingsFabric/issues/10;

* Create gradle.yml

* Fix workflows/gradle;

* Add: 1.18 support

* Fix two deprecations

* Remove workaround Mixin

* Fix Workflow

* Make Gradle wrapper executable

* Resolve conversations

* Resolve conversation

* Setup GitHub Actions;

* Fix Cache problem in Actions;

* Add Gradle Wrapper validation;

* Update gradle.yml

* Fix calculation of sound fx in water (gain calculation for low pass data);
https://github.com/OreCruncher/DynamicSurroundingsFabric/issues/12;

* Fix: https://github.com/OreCruncher/DynamicSurroundingsFabric/pull/17#issuecomment-1014811491;
Add: fading water ripple at the end;

* Edit: fading water ripple at the end;

* Add support for 1.18.2;

---------

+++ src/main/java/org/orecruncher/dsurround/config/WaterRippleStyle.java
@@ -15,7 +15,7 @@ public enum WaterRippleStyle {
 
         @Override
         public float getU1(final int age) {
-            return (age / 2) * this.DELTA;
+            return (int) (age / 2F) * this.DELTA;
         }
 
         @Override
